Introduce a tokenization step
Properly dealing with macros requires the ability to deal with the
arguments as tokens. Naive parsing cannot handle cases like
`foo!( ")" )` as it will not know that the inner right-paren is inside
of a string and does not close the macro invocation.

The tokenizer is pretty dumb at the moment. Our tests continue to
pass, but the overall pass rate across a wide set of crates has
decreased. Some of this is due to missing tokens, but a better
tokenizer will require porting the smarts to deal with the various
types of strings, comments, numbers, etc.

Hopefully, the entire parser can be moved to the tokenizer, avoiding
maintaining two parallel similar bits of code.